# static/openapi.yaml
openapi: 3.1.0
info:
  title: Spotify GPT Agent
  version: 1.0.0
servers:
  - url: https://spottyg.onrender.com
paths:
  /search_track:
    get:
      operationId: searchTrack
      description: >
        Searches for a track in the Spotify catalog. 
        This endpoint requires a bearer token from the token handoff process.
        Authenticate by visiting https://spottyg.onrender.com/login,
        then retrieve your token at https://spottyg.onrender.com/token_handoff.
        Paste the token into the GPT when prompted.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Track search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                  error:
                    type: string
        '500':
          description: Error searching track
  /search_album:
    get:
      operationId: searchAlbum
      description: >
        Searches for an album in the Spotify catalog. 
        This endpoint requires a bearer token from the token handoff process.
        Authenticate by visiting https://spottyg.onrender.com/login,
        then retrieve your token at https://spottyg.onrender.com/token_handoff.
        Paste the token into the GPT when prompted.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Album search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                  error:
                    type: string
        '500':
          description: Error searching album
  /get_playlists:
    get:
      operationId: getPlaylists
      description: >
        Retrieves the user's playlists from their Spotify account. 
        This endpoint requires a bearer token from the token handoff process.
        Authenticate by visiting https://spottyg.onrender.com/login,
        then retrieve your token at https://spottyg.onrender.com/token_handoff.
        Paste the token into the GPT when prompted.
      responses:
        '200':
          description: User playlists
          content:
            application/json:
              schema:
                type: object
                properties:
                  playlists:
                    type: array
                    items:
                      type: object
                  error:
                    type: string
        '500':
          description: Error retrieving playlists
  /create_playlist:
    post:
      operationId: createPlaylist
      description: >
        Creates a new playlist in the user's Spotify account. 
        This endpoint requires a bearer token from the token handoff process.
        Authenticate by visiting https://spottyg.onrender.com/login,
        then retrieve your token at https://spottyg.onrender.com/token_handoff.
        Paste the token into the GPT when prompted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                public:
                  type: boolean
                collaborative:
                  type: boolean
      responses:
        '200':
          description: Playlist created
        '500':
          description: Error creating playlist
  /add_to_playlist:
    post:
      operationId: addToPlaylist
      description: >
        Adds tracks to an existing playlist in the user's Spotify account. 
        This endpoint requires a bearer token from the token handoff process.
        Authenticate by visiting https://spottyg.onrender.com/login,
        then retrieve your token at https://spottyg.onrender.com/token_handoff.
        Paste the token into the GPT when prompted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                playlist_id:
                  type: string
                track_uris:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Tracks added to playlist
        '500':
          description: Error adding to playlist